Extract shared permission definitions
Move ALL_PERMISSIONS and VALID_PERMISSIONS into a standalone
module without server-only so both server code and client
components import from the same source. This eliminates the
duplicated PERMISSION_GROUPS constant in role-form-dialog
that could drift out of sync.

Also fix withTransaction mock callback type that caused
TS2345 under tsconfig.typecheck.json.

Refs #134
