+++ src/lib/auth/permissions.ts
@@ -35,31 +35,9 @@ async function loadRolePermissions(roleName: string): Promise<Set<string>> {
 
 // ── Constants ─────────────────────────────────────────────────
 
-/**
- * Single source of truth for all permission strings in the system.
- * Used for the permission checkbox grid in role management UI and
- * for validating permission values on the server side.
- *
- * Grouped by resource for display purposes.
- */
-export const ALL_PERMISSIONS = {
-  accounts: ["accounts:read", "accounts:write", "accounts:delete"],
-  roles: ["roles:read", "roles:write", "roles:delete"],
-  customers: [
-    "customers:read",
-    "customers:write",
-    "customers:delete",
-    "customers:access-all",
-  ],
-  "audit-logs": ["audit-logs:read"],
-  dashboard: ["dashboard:read", "dashboard:write"],
-  "system-settings": ["system-settings:read", "system-settings:write"],
-} as const;
-
-/** Flat set of all valid permission strings. */
-export const VALID_PERMISSIONS: Set<string> = new Set(
-  Object.values(ALL_PERMISSIONS).flat(),
-);
+// Re-export from the shared (non-server-only) module so that existing
+// server-side consumers can keep importing from this file.
+export { ALL_PERMISSIONS, VALID_PERMISSIONS } from "./permission-defs";
 
 // ── Public API ─────────────────────────────────────────────────
